requirements.txt
amqp==5.0.6
appnope==0.1.2
asgiref==3.4.1
astroid==2.7.3
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
Django==4.0.4
django-allauth==0.45.0
django-bootstrap-form==3.4
django-crontab==0.7.1
django-extensions==3.1.3
django-filter==2.4.0
django-forms-bootstrap==3.1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
flower==1.0.0
humanize==3.11.0
idna==3.2
inflection==0.5.1
ipython==7.26.0
ipython-genutils==0.2.0
isort==5.9.3
itypes==1.2.0
jedi==0.18.0
Jinja2==3.1.2
kombu==5.1.0
lazy-object-proxy==1.6.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.2
mccabe==0.6.1
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.3.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycodestyle==2.8.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
pylint==2.10.2
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
uritemplate==4.1.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
